Use real data for the dashboard graphs
* Objective

This commit objective is about using real data for the dashboard graphs.

* Why

Using mock data was ok, but at some point, we needed to use the data
coming from the APi.

* How

This commit updates the Charts so they start using real data instead of
the mock data we were previously using.