Use real error-rate data
* Objective

This commit objective is about using accurate error rate data on the error
rate screen.

* Why

We were using actual data for the rest of the charts in the dashboard, but
for the error rate, the error rate currently has some issues as the
source that we've selected in the backend doesn't fail, so we have
made a few tricks on the backend to simulate some error rate.

* How

This commit starts using the error rate received from the backend.