Add Program Performance & Throughput and Refactor
* Objective

This commit has two separate objectives:

1. Adds the Program's Performance details and Throughput charts on the
   Transactions Page.

2. Perform a small refactor to perform the HTTP calls on the Pages
   instead of each component.

* Why

We were missing to add Program performance and throughput charts for the
Programs in the Transactions table.

We were also making tons of API calls (one per chart component)
that could hinder the performance of the API.

* How

This commit adds the Program's performance and Throughput charts.

This commit performs the refactor of making the API calls happen on
the different pages.