Closed version of encinitas
* Objective

This commit objective is about creating a closed version for encinitas.

* Why

The servers for encinitas will be down for some time so this is meant to
be the closed status.

* How

This commit disabled the demo in the Home, and make the app to always
render the Home in the Dashboard o Transactions